# app/openapi.yml
openapi: "3.0.0"
servers:
  - description:  Test
    url: /
info:
  description: This is a simple API
  version: "1.0.0"
  title: Test API
paths:
  /car:
    post:
      tags:
        - Car
      summary: Add a new car
      operationId: app.views.carView.add_car
      description: |
        Add a new car
      requestBody:
        description: Creates a new Car 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Car
      summary: List all cars 
      operationId: app.views.carView.get_car
      description: |
        List all cars
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllCarResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /car/{car_serial_no}:
    get:
      tags:
        - Car
      summary: Find Car by car_serial_no 
      operationId: app.views.carView.get_car_serial_no
      description: |
        Returns a single car 
      parameters:
        - name: car_serial_no
          in: path 
          description: Car serial Number 
          required: true 
          schema:
            type: string 
            example: C-1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCarResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      tags:
        - Car
      summary: Update/Patch Car by car_serial_no 
      operationId: app.views.carView.update_car_record
      description: |
        Updates a single car 
      parameters:
        - name: car_serial_no
          in: path 
          description: Car serial Number to update
          required: true 
          schema:
            type: string 
            example: C-1
      requestBody:
        description: Updates a car record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Car
      summary: Deletes Car by car_serial_no 
      operationId: app.views.carView.delete_car
      description: |
        Deletes a single car 
      parameters:
        - name: car_serial_no
          in: path 
          description: Car serial Number to delete
          required: true 
          schema:
            type: string 
            example: C-1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /customer:
    post:
      tags:
        - Customer
      summary: Add a new Customer
      operationId: app.views.customerView.add_customer
      description: |
        Add a new customer
      requestBody:
        description: Creates a new Customer 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Customer
      summary: List all customers 
      operationId: app.views.customerView.get_customer
      description: |
        List all customers
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllCustomerResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /customer/{customer_id}:
    get:
      tags:
        - Customer
      summary: Find Customer by customer_id 
      operationId: app.views.customerView.get_customer_by_id
      description: |
        Returns a single customer 
      parameters:
        - name: customer_id
          in: path 
          description: Customer ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCustomerResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      tags:
        - Customer
      summary: Update/Patch Customer by customer ID 
      operationId: app.views.customerView.update_customer
      description: |
        Updates a single customer 
      parameters:
        - name: customer_id
          in: path 
          description: Customer ID to update
          required: true 
          schema:
            type: string 
            example: 1
      requestBody:
        description: Updates a customer record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Customer
      summary: Deletes Customer by customer ID 
      operationId: app.views.customerView.delete_customer
      description: |
        Deletes a single customer 
      parameters:
        - name: customer_id
          in: path 
          description: Customer ID to delete
          required: true 
          schema:
            type: string 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
    Car:
      required: 
        - car_serial_no
        - price 
      type: object
      properties:
        car_serial_no:
          type: string
          example: C-1
        make:
          type: string 
          example: Honda
        model:
          type: string 
          example: CBR1000
        year:
          type: string
          example: 2020-04-04
        color:
          type: string 
          example: blue 
        price:
          type: number
          example: 1050000.0
    CarPatch:
      type: object
      properties:
        make:
          type: string 
          example: Honda
        model:
          type: string 
          example: CBR1000
        year:
          type: string
          example: 2020-04-04
        color:
          type: string 
          example: blue 
        price:
          type: number
          example: 1050000.0
    GetAllCarResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/CarObject'
        status:
          type: string 
    GetCarResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/CarObject'
        status:
          type: string 
          example: success
    CarObject:
      type: object
      properties:
        car_id:
          type: integer
          example: 1
        car_serial_no: 
          type: string 
          example: C-1
        color:
          type: string 
          example: blue 
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
        make:
          type: string 
          example: Honda
        model:
          type: string 
          example: CBR1000
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
        price: 
          type: number
          example: 1050000.0
        year:
          type: string 
          example: 2020-04-04
    Customer:
      required: 
        - phone
        - email_id
        - ssn
        - first_name
        - last_name
        - address 
      type: object
      properties:
        phone:
          type: string
          example: '9999999997'
        email_id:
          type: string 
          example: cust_1@example.com
        ssn:
          type: string 
          example: '1234'
        first_name:
          type: string
          example: John
        last_name:
          type: string 
          example: Doe 
        address:
          type: string
          example: Abc Apartments, CA-94301
    GetAllCustomerResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/CustomerObject'
        status:
          type: string 
    CustomerObject:
      type: object
      properties:
        address:
          type: string
          example: Abc Apartments, CA-94301
        customer_id:
          type: integer
          example: 1
        email_id:
          type: string 
          example: cust_1@example.com
        first_name:
          type: string
          example: John
        last_name:
          type: string 
          example: Doe 
        phone:
          type: string
          example: '9999999997'
        ssn:
          type: string 
          example: '1234'
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetCustomerResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/CustomerObject'
        status:
          type: string 
          example: success
    CustomerPatch:
      type: object
      properties:
        address:
          type: string
          example: Abc Apartments, CA-94301
        email_id:
          type: string 
          example: cust_1@example.com
        first_name:
          type: string
          example: John
        last_name:
          type: string 
          example: Doe 
        phone:
          type: string
          example: '9999999997'
        ssn:
          type: string 
          example: '1234' 
...
    ApiResponse:
        type: object
        properties:
          msg:
            type: string
          status:
            type: string 